+++ src/components/feature/newArrivals/EventBanner.tsx
