+++ src/components/feature/recommendations/EventBanner.tsx
